+++ packages/walrus-site-builder/CHANGELOG.md
@@ -1,6 +1,6 @@
 # @cmdoss/walrus-site-builder
 
-## 3.0.0
+## 2.6.0
 
 ### Minor Changes